Solved decimal axis problem
Now all of the axis are integers, and share the same distribution as the ones from the other graph. New issue arose, very tiny image when long time passes. I need to save the image as the individual set it up to be, so after the changes he/she wants to make. A way to resolved this is adding a save button.